Finished day 6
This was interesting because the code was close, but the brute force method
did not work. I needed to use a ray-tracing idea to process one light at a
time. That worked wonderfully. Sadly, I think I can make the code a lot
smaller but then it'd be more complex to handle the boolean vs. analog
lights.