Finished Day 15
This was fun because of the desire to be able to create a generalized
function that would contain a certain number of integers, and sum to a given
value. This was a key part of the solution, and it was interesting to use
the recursion - but not `loop` in the code.