Got Day 21 part 1
The first part seemed not too bad, and the second part seems reasonable, but
while I can't seem to find a looser that is the right answer - it's too low.
That means I need to buy more, but not get more. And I've been over the
prices, and I can't see how to do that.