Finished Day 21
This one was no fun because it was a pricing problem, and yet I had no idea if
it was in the win/lose logic or the pricing, and getting the first part did
nothing to help me get the second. But after looking at the mixtures again,
I saw the two-ring solution, and that was what effected my pricing table.