Set everything up for 2016
I didn't know what I was thining originally, but with the new problems
for 2016, it was clear that I needed to put in a little more structure
for the problems this year. That's all this is - making the year a part
of the namespacing, and the source data, and then off we go for 2016!