Finished Day 7
This was a tricky one because the data wasn't represented in the
examples, and I had to go and see Bret's solution and then see that the
data was and see the multiple patterns in the data. This was then a bit
of a pain for the second part, but it's still just pattern matching.