Finished Day 3
This was kinda a pain... I didn't want to go all the way to having the
state of the entire cloth piece, but try and I might, I could not see a
way to calculate the union of all the intersections without something
like this. The problem was in the multiple-counted areas, and removing
them from the total - it's not an easy problem.