Finished Day 15 - 2016
This was fun because I wanted to try to do it with infinite sequences of
the positions of the discs, and then offset them to remove the time of
travel on the machine - thus looking for a sequence of zeros - meaning
that the tumblers lined up, and there was safe passage through the
machine. This was maybe not the most efficient, but it was a lot of fun.