Finished Day 4
This was a great bit of fun to do! The data needed to really be fixed
up, but once tou got the data in the right format, the code almost wrote
itself. I was able to refactor a lot of the code when I saw the second
part, and that made the solutions really collapse. What a hoot! This is
just what I was hoping for! :)