Partial Day 19 - 2016
I was abkle to brute force part 1 in no time, and even had a brute force
scheme for part 2, but it didn't check with the site - so I looked and
just can't find the problem or a simpler way to solve the puzzle. So
this is going to be use Part 1.