Finished Day 6
This was interesting because I took Bret's idea of using the locations
as keys in a map, and not having to actually 'paint' anything. This was
a nice change to the old scheme. Also, I really liked seeing that the
points to exclude in the first part were those that were 'closest' to
the points on the perimeter. Very nice! :)