Finished Day 21 - 2016 - Part 1
I was able to again parse the instructions into a useful set of
functions and arguments, and then the sequence, with a loop/recur, was
able to process the password. That's just fine. But then they want to
reverse it, and honestly, I just didn't have the interest at this point
to do part 2.