Finished Day 7
Wow... this was not fun - the graph didn't help, so I used a simple set
of dependencies map - like a Makefile, and then ran through a series of
very simple functions to check the status of the running steps, and what
needs to be done. It was a lot of messing around.