Finished Day 8
This was interesting because I was able to make an `eat` functon - like
`take`, but it removed the elements from the LinkedList source so that
they would not be processed again. This was the key to using recursion
to build the data structure, and then the counting was not bad at all.