Finished Day 10
This was annoying in that once again it wasn't about the coding, it was
about an algorithm to find the data to look at. I understand that this
is meant to be challenging, but at this point, these are not as fun as
the ones where it's up to us to come up with the clever solutions.