Fixed the terminal test on Day 9
I had a bad terminal test on Day 9, and didn't catch it until I was
going back through all the different test cases and found one that was a
multiple of 23, and so I was missing the last points, and that was
throwing off the total. Simple fix once I was on to the probelm.