Finished Day 11
This was fun because I completely refactored the problem when I got to
Part 2 as it was going to take way too long to search for Part 2. And
the refactoring was so nice as it was all about `partition` and how I
could be processing an entire row at a time, and then a complete
sub-grid... and then just assigning the location values and finding the
max. It was a lot better solution that I originally had.