Finished Day 12
This is another annoying one - and I'm going to have to remember this -
if the numbers look impossible - then they DO NOT expect you to actually
use a machine to solve this. You have to print out the evolution and
then see the pattern, and when I did that, Part 2 was simple. But it's
annoying.