Finished Day 13
This was a lot of work because the input was a track we had to mo ve
along. Keeping track of the carts on the track, and then manipulating
them as the rules indicated. It was a lot of work to get it to match the
answer - even when the tests were fine. Not easy.