Finished Day 18
This was a classic 2D Game of Life, and the set-up for the second part
was that depending on the starting point, there is usually a pattern
that emerges in Life, and we just had to develop it, find the cycle, and
then make part 2 the mapping of the generations into that cycle. Not too
bad given that it was clear they didn't expect it to run all that time.