Have a workable solution for Day 17
This is now a workable solution for Day 17 - there are some efficiency
issues that could be addressed, and the real problem is that while the
solution appears to be right on the money, they AoC site says it's too
high. This is because there is some rule I'm not correctly implementing,
but this solution is so much simpler than the previous one, it's hard to
imagine what it is. Again... the solution looks correct, and stable. So
this is likely to never be really solved.