Finished Day 19
This took some help - like as in disassembling the code to see what it
was doing, and that wasn't as fun as it sounds. There were no patterns,
and there were no real things to look at - save the registers and
stepping through the code. The answer wasn't all that fun - but I now
know to do this, if it comes up.