Finished Day 21 - Part 1
I was able to get part 1 done, but finding the pattern in the sequence
was just too much... I looked for repeats, and that wasn't it... and
then the one *before* the first repeat, and it didn't seem to hit. So I
just have to give it a pass for today... I missed.