First Part Day 23
The first part of today was not bad, but the problem with the second
part is that it's just impossible to search every point, and it's not
any easier to 'paint' the space for each radius. That means that there
needs to be something else, and I've seen folks talking about using
libraries to help them. Well... yippee... but that's not working here.