Start of Day 24
I finally got the example to work - and the trick was to update the
units DURING the attack - and that means the 'damage' is updated at the
time of the attack as well. That, and the proper ordering of the
attacks, and we got their answer. But then when applying it to the
complete puzzle, it says we're low at 21378 - but it's got to be less
than the starting size of all infection groups - 24087. So there is
something close, and I haven't found it yet.