Finished Day 23
Part 2 was a binary search - on the volume, and not on points. The key
was to get it, by volume, down to 32 points, and then do an exhaustive
search from there. This then allows us to return the top-10 in terms of
counts and the distance. Take the first one, and we have a winner. I'm
sure there are issues with this, but then that's OK... I know where they
would be.