Refactored to add util and path
We really needed to have a few more namespaces to make things a little
more manageable, and this represents that initial refactoring. I was
hoping to have a nice way to do A* pathing, but this is turning out to
be a little more difficult than I'd hoped. So I wanted to get this in
before I went too far.