Finished Day 7
This was interesting because it was first about creating a decent data
structure that would allow us to find out the data we needed. Then it
was pretty simple to see what needed to be corrected, and by changing
that, the tree balanced. But then I went on to make it a lot easier to
spot the issue for another run - just to be nice.