Finished Day 11
This was interesting because we added to the util namespace with `sign`,
and we created a new namespace for dealing with hex grids. I can just
see this coming back again and again in these puzzles, so lets break it
out now, and see what we can do with it. Very nice puzzle today.