Finished Day 12
This was interesting because I had to noodle on the best way to solve
this for a while before I came up with the structure for `talkers`. Once
I had that, the rest was easy, but it was really the loop in `talkers`
that took the effort. But good fun.