Finished Day 15
This was interesting, because I went the sequence route again, and the
time just wasn't there. So I fell back to the simmple iterator and a
doseq. This was plenty fast, and pretty simple. Then on part 2, it was
simple to put a loop in each generator's calculation, and that gave me
all the answers I needed.