Finished Day 16
This was an interesting simulation again, this time, parsing the dance
moves into maps of functions with arguments, and then running them in
the right sequence. And part 2 was all about running enough of them to
find a pattern, and sure enough... there was one - and we then just had
to codify that pattern, and we're good to go.