Finished Day 17
This wasn't too bad on part 1, but part 2 was all about seeing a pattern
in the values that were placed after 0 in the sequence. This could be
seen as a rotating queue, and so looking for those that hit the first
space, and then saving only that one. I needed help on part 2.