Finished Day 18
This was really interesting because I had the solution - but with one
mistake on the `snd` - I was using `conj`, and the sequence wasn't a
vector any longer - after the `rest`, and that meant that I was putting
the numbers on the wrong end. Once I saw that it was all easy.