Finished Day 21
This was interesting because the flipping and rotating was a lot more
than the puzzle indicated - it was *all* possible combinations, and then
I used `distinct` to make it a little more reasonable. The second point
was that for Part 2, I had a problem in the row/col switching, and I
needed to fix that, and then we got the right answer. But it worked fine
for Part 1 - which was odd.