Finished Day 22
This was interesting because I happened to just have a lot of the
typical 'paining' routines for the grid, and all I needed to do was to
look at the other 'traveling' puzzles in the series for turns and state
change, and piece it together. Then I refactored it a bit to make one
`move` function, where you could pass in an optional state-change
function to determine how the virus carrier changed the state of the
nodes. This was all that differentiated the two parts.