Finished Day 23
This was another machine simulation where we had to disassemble the code
to see what it's trying to do. And I got to the point that I actually
got the point, and then checked to see that I was right - and I was off
by one, but that's OK. This was one of those class of problems that
compute strength doesn't help.