Moved to clojure.java-time from Joda
The Joda time folks are saying that they are asking people to move to
JDK 8's date objects, and that means that we need to do Java interop, or
we need to use something like clojure.java-time. So this change is just
that move, and thankfully, there isn't a lot to change, but it's nice to
get on something that is going to be supported going forward.