Finished Day 2
This was fun, and it was another of the computer simulations based on
integers this time, as opposed to text-base opcodes. So it's just a
little easier to work with, but it's all the same, really. It was a
little odd to figure out the indexes, but after a few trys, I got what
they meant by the replacing 12 and 2... then it wasn't bad at all.