Switched to loop/recur
I know the recursive call worked on 'step', but it wasn't what I really
wanted, and it wasn't as simple as changing that to a 'recur', so I just
took the time to make it loop/recur on the function, and then it wasn't
just a step - it was a complete 'run', and then the noun/verb call was a
run with some args. It was just a little work to make it nicer.