Finished Day 5
This was a little more tim-consuming because they really added in
several opcodes, and access schemes. Still, I wanted to make it as clean
and concise as possible. So I didn't just copy/paste code, and change
one thing, I made it fewer lines of code, and that's my goal.