Finished Day 6
This was fun in that it really was something I could have seen from the
start, a typical tree problem, but I didn't get carried away with the
first part, as that was a simple counter - and that set me up for a nice
and easy second part. The matching up of the climbing patterns just came
to me, and then it was all pretty easy. Nice. :)