Added I/O pausing to run function
In order to handle Day 7, we needed to handle I/O pausing of the
execution. The easiest way to do that was to make the complete machine
state a map, and then allow for either the old-style args, or the new
state map, and then update the state on output - consuming inputs, and
appending outputs. The pausing for I/O was needed due to Day 7, part 2,
where the output of one program had to be fed into the next amplifier as
input - but without losing where each program was. Not bad... :)