Finished Day 9
I had to go back and add a new addressing mode to the Intcode computer,
and at the same time, realize that there is a simpler way to handle the
existing addressing modes, and that's all good news. Then I had to add
in the other properties - unlimited size, and that wasn't all that bad,
either. In all, it was a fun day.