Finished Day 10
This was interesting because the first part was just trying all the
different angles to the target, for each point, and seeing how many
unique angles there were. That's not bad. Then the second part was
clever in that if we add 360 to each of the colinear points, as they
radiate from the platform, then we can then sort them all by the angle,
and this puts everything into the right order. Then it's just taking the
right number from the sequence. Done.