Update the log message
As I'm doing some logging, the new nREPL thread name is a UUID, and I
find that exceptionally long, so I updated the logging message format to
just keep the last 12 digits of it, and then we'll see how that goes.
It's not something I'm that worried about, but it's annoying to be too
long. Also, I removed the session ID and caller ID from the format,
because we don't need them in this project.