Moved to longs for the Intcode programs
As it turns out, on Day 11, I found that the Intcode data really needed
to be parsed as `long` values, and not `int`. So I needed to come back
into these days, and update them to use the right parsing function.
Nothing else changed, and these days weren't effected, but I just wanted
to make sure.