Fixed the state management
Two things I messed up the first time - the first was that the Intcode
data needed to be parsed as `long` values and not `int` -- but it wasn't
until Day 11 that I had data that really pointed this out. Wild - all
the previous tests and days passed. OK... the second one was that the
relative base offset needed to be persisted from one invocation to the
other - to properly maintain state. This was a silly mistake on my part,
but all the examples to date never revealed it. Still... I also cleaned
up a little shite space, and added the ability to log things as that is
far more helpful most of the time.