Finished Day 11
Well, this was interesting... I found that I had two issues in the
previous Intcode computer work - I wasn't parsing the data as `long`
values, and then I wasn't properly maintaining the relative base offset
from one run to the next. It just never mattered to any of the previous
tests or runs. But when I got those fixed, I was able to finish today's
with a simple paint program and then a function to take the painted
tiles, and render them into a reasonable image.