Added repl.it Support
The repl.it folks have made it possible to load projects from GitHub,
and in doing that, they can support a full Clojure dev environment. So I
decided to update the project with the suport file, and a badge in the
README.md so that it's clear how to get this project up and going. Nice.