Added factoring and lcm
In Day 12 I needed to have a prime factorization function as well as a
least-common multiple routine. I saw these in some other libraries, but
they didn't allow for multiple arguments, so I looked up the algos, and
wrote them for the project. Very nicely done! I have to say.