Finished Day 12
This was a blast - as the first part is pretty clear what to do, and it
doesn't take too long - but part 2 is clearly not something you can
brute force an answer. So the key is to look at the period of the
individual components - x, y, z - and then look for the LCM of their
periods to come up with a time when they all will repeat. Still not
trivial, but it's a lot more solvable.