Finished Day 13
This was a wild one... the first part wasn't bad at all, but the second
part had me stumped for a long time. First, I just tried to manually set
a series of paddle inputs, and that worked for a while, but I felt there
had to be an algorithm for the paddle inputs, and so I tried to plot the
trajectory of the ball, and see where it intersected with the plane of
the paddle. Missed. Then I decided to track all the bounces - still
missed. Finally, I looked and saw that a simple call to the compare
function on the x-axis values of the ball and paddle worked, and the
solution fell out. Wow...