Finished Day 14
This was nasty - very... I tried to do the mapping, and that worked for
all the trial data, but not for the real data. And then I switched to a
more incremental way, and that was fine for part 1, but part 2 was too
long - so I was extrapolating, and then found a simple checker for the
answer, and that got me. But I really think the answer was to finish the
initial mapping, but that just didn't work. Shucks.