Finished Day 16
This was another interesting day where the first part was fine - but the
second was computationally impossible if you tried to just run it on the
data. Rather, what I did was to see that for the latter half of the
data, the phase function was just summing the remaining digits in the
sequence, and since the offsets were taking more than half of the data,
then it was clear that I just needed to make an efficient function for
the summing of the digits, and that was very easily done. Very slick.