Day 15 is a Mess
Just by looking at this, I know I need a painting algorithm to "find the
perimeter without regard for the shape. That's something I can look up,
and it's not horrible, but it's not a lot of fun. And then we're back to
a path-finding algorithm, and I just don't have those in my tool belt
yet, so I'll hold off on this day for a little while, and then come back
to it in pieces... first the painting to find the oxygen sensor...