Finished Day 15
OK... it hit me - I didn't have to paint the area - I could recursively
do it by copying the CPU for each of the four possible directions, and
then, for each successful move, recursively call the same function. This
was very fast, and exposed the entire space I needed to find. Very nice.
And the second case was a simple diffusion problem, and that wasn't at
all bad.