Finished Day 17
This was interesting with the Intcode computer again, and the first part
wasn't too hard, but the second was a mess for a wihle, as I had the map
rotated, and that messed up everything. When I realized that, it still
took a few minutest to anually solve the second part, but it wasn't too
bad.