Finished Day 19
This was interesting in that the initial brute force was again easy for
part 1, but there was an obvious simplification for part 2 - as the beam
spreads, we really only need to calculate the points that are likely in
the beam - and a few to either side. That means we change the storate to
a very sparse mode, and that made for a much simpler solution.