Start on Day 18
I have a decent start to part 1 - but we're getting into search-space
trouble for trial #4, and the puzzle. It's just too big, and it takes
too long to find the optimimum path for all the keys. Again, it might
be the number of keys to grab... but something has to be done to reduce
the search space.