Upped the memory and stack
Some of the tests for Day 20 were blowing out the stack, and so I wanted
to make sure that there wasn't going to be a problem with the memory or
the stack, so I added a few JVM options to the config. Nothing big - but
more than the defaults to make it a little roomier.