The Logic Works, but too complex
This is a nasty one because the most significant issue was in parsing
the input data to get the portal endpoints -- and it was a serious pain.
But then it's a simple matter of exploring to find the shortest path,
and we have that from Day 18, or so. Problem is... the size of the maze
is too big to be solved by this 'explore' function. Which is why we had
to increase the stack space on the JVM - as it was overflowing. Not that
the logic is *wrong* - but it's just not efficient enough for this sized
puzzle.