Finished Day 23
This was interesting because there was a lot of state to manage. I'm
sure I could have done it a lot differently, had I know I would be doing
"complete passes" on the machines up-front, but it still works just
fine. I'm not a huge fan of these state problems, as they really don't
make for clean solutions.