Finished Day 22
This was something I really had to look into - because there were a few
really tricky points in the math that I just didn't know. First, that
there really is a 'modinv' function for coprimes, and that makes sense
as the examples are 1:1 mappings - so once I found that, and how to
use the Extended Euclidean algorithm, I could code that up. That made
getting one, two, even 70,000 reverse shuffles possible. But that wasn't
enough. The next biggie was realizing this reverse shuffle was linear -
and I did not see that. But if it is, then you can fit an A*x + B
through it, and then you can repeat this with a polynomial expansion and
get the answer. It's crazy because again, you need to be able to use a
modpow function, that's on BigInteger, but I never would have thought
that modulo was linear.