Finished Day 18
This was a very interesting day. The permutations worked just fine - as
long as the key set was small, but even on some of the examples, it
wasn't small enough. Then I went with caching the searches, but that
didn't work out when you have a _specific_ search set. What worked was
to have a point, and a set of keys to find, and that caching worked
perfectly. So it was very much like a second-order path finding, and
that was a lot of fun.