Added Day 18-like logic
I was thinking that we could do the same kind of thing on the pathing as
Day 18 - which worked out very well indeed. The problem I ran into is
that even getting the distances between portal openings is proving to be
computationally costly, so I don't know what I need to try next.