More work on Day 20 part 2
This works, but it's not performant enough, which means it's really
"wrong" for the problem. I've looked at several possible solutions, and
they seem to go back to the simple path-finding, but I know that's not
going to work -- so I need to have the things I have... but I need to
somehow trim the search space, or do a different path-finding once I
have the inter-portal distances.