Finished Day 20
Finally got the idea of the BFS - as opposed to Dijkstra's - for a
distance determination. The key is that it's not reentrant, but just a
simple looping construct - that stops at the FIRST encounter with the
target point. Since each possible path is moved one step at a time, and
all the points are marked in the visited set, then the first one to the
target wins. Always. This was a beautiful solution, but I still needed
to use the warping logic from before, and then it was easy to make it
all come together.