Added acl and user creation
For the future, I needed to update the packages installed to include the
'acl' package so that 'setfacl -Rbn workspaces' could remove all the
advanced permissions that I didn't want on the directory. At the same
time, I will want to persomalize things with a dotfiles repo in GitHub
so that the files there can be placed on all new Codespaces, and in
order to do that, I wanted to put in the section about making my user.
It's not enough, but it's a start.