Fixing up to run cleaner in Codespaces
Woth my dotfiles worked out with CQ, and that included my ~/.lein we
don't have to include the plugins here for the project, as they will be
picked up in the dotfiles. This is a really sweet update as it means I
get all the same dev tools with dotfiles and my own account.