Finished Day 2
This was interesting in that I needed to code up an 'xor' function for
an arbitrary number of arguments, and I can see why a macro would be
good in this case, but the one I looked at didn't work as I needed it
to, so I went with a simple function instead. It still short-circuits on
the second truthy value, but it isn't as efficient as a macro might be.