Finished Day 7
Part 2 was interesting as it was a search that needed to be done
recursively and keep the data clear all the way down and back up. So
it's something that was a little tricky to get right, and took some time
playing with the data structures, but we got it.