Finished Day 8
This was another fun one - I enjoy the CPU simulations as it's fun to
parse the source code, and then interpret the instructions, and in this
case, I didn't have to really leak too much of the virtual machine to
get the answers. Just make sure to preserve the accumulator and PC, and
we're in good shape. Fun day.