Finished Day 10
This one was OK for the first part, but the second part got ugly - fast,
as I could not figure out a simple way to accurately count the possible
drops. WHat I came up with wors, by looking for single- and double-
drops - independently, and summing them, but it's not elegant at all.
but by breaking up the total sequence on a difference of 3, I was able
to make it a tractable problem by computing the counts on each smaller
section, and multiplying the together.