Finished Day 14 - Part 1
This computer simulation isn't horrible - and part 1 wasn't too bad, but
part 2 is going to take a little time to figure out how to run through
all the different values of the 'X's in the mask. I'm guessing it's
something to pre-process in the parsing, but maybe that's a waste as
these aren't run more than once.