Finished Day 13 - Part 2
This one I needed help on, and it was based on the Chinese Remainder,
that Bret remembered. This was something I really need to read up on,
but it's amazing how the LCM was in the answer, and because they were
all primes, it was a simple multiplication of them. This was hard.