Finished Day 16
I had a problem on Part 2 because I was removing the bad tickets, but I
wasn't then *using& the reduced set, and that was a big problem. When I
fixed that, the rest fell into place. Silly mistake, but I was tired,
and it was at the end of the day.