Finished Day 17
What was tricky about today's puzzle - for me - was that the example
appeared to be aligned, and it wasn't. Just the minimal node
representation. And that wasn't clear until I looked at the pattern of
the nodes. Then it was clear. But going from 3D to 4D was just a repeat
of the problem, and it wasn't worth trying to find a simpler way to
collapse the code into a general solution.