Finished Day 18 - Part 1
This wasn't all that fun, but I was able to get it because it was all
pretty linear, and the parens just added a level to the calculations. As
I tried to do Part 2, I knew the answer was to reorganize the data into
a tree, but I just didn't have the interest. I'll have to see if I get
back to it later in the month.