Finished Day 20 - Part 1
This wasn't all that fun because it's just another brute-force issue -
you run all the data, and you get the edge matchings that you need. It's
just that I wanted to see what Part 2 was, and then see if it was
something to continue. I'm not sure... but I can do it later.