Day 21 - Parsed Input
This one makes no sense to me or Bret, but I have the parsing done, so
I'm going to check it in, and we'll see if I get back to it. I think
there's something in the language of the problem statement that I'm
missing - something that would point to a type of partial-information
problem. But I'm missing it.