Finished Day 23 - Part 1
This wasn't too bad, it was just a bunch of data manipulation, until we
got to part 2 - and then it was again, one of finding a pattern to the
updates, and being able to create a formula for that, as it is clearly
too computationally costly to just run the large list that many times
through the process. Maybe I'll come back to it.