Finished Day 18 - Part 2
This was silly... I had it - I really did - but I was using parse-int on
the numbers, and some of these blew past that, and it was truncating the
longs to ints, and that was the problem. When I just went to parse-long,
I was good to go. Logic was great - I missed an easy one.