Finished Day 6
This was funny, because my first approach was to keep track of each fish
individualy, but that leads to a computation problem as it just doesn't
fit into any reasonable machine... then I tried to plot it and
interpolate on a log-linear graph, and it was close, but too imprecise.
The solution was to group the fish into ages, and then treat them as
one. The updating logic is a little more, but it's still not bad, and
the values change, but the computation size is exactly the same for any
sized input - as the ages are a fixed set. Nice.