Finished Day 7
This was fun because it really was clear that calculating the complete
function looking for the minimum was the wrong thing to do - and
understanding the median and mean were a lot better choices. This was
fun to pull in those functions, which I've now added to the util
namespace.