Cleaned up Day 8
There was no value in parsing the data into a map - we can make the map
from the sequence of strings, and that's good enough. It also makes the
parsing a little simpler, as the same function is done to both parts of
the line. It's no different, just a little cleaner.