Finished Day 9
This was interesting because it's alomst as easy to just confider the
four directions, as it is to make a loop to consider the four
directions. It's just that in the first part, there's no value in the
loop, but in the second, the path finding - to hit a '9', it pays off
because of possible back-tracking. So easy enough, and it's all good.