Finished Day 10
This was fun because it was a nice reminder of a simple stack operation,
adn with the looping, it was not bad at all. The combining of the two
outputs in the same function was simply for debugging Part 1, but it
really came in handy with the 'mate' function to solve part 2. Nice fun.