Finished Day 11
This was fun because the individual step was an interesting iterative
breadth-first problem that was surprisingly tricky to get. I was hoping
to have a single inequality on the flashing, but it really was far
easier to break it into two as the steps could be more than 1, and that
made it interesting.