Finished Day 13
This was interesting, and I should have gone ahead and made the
visualization function for part 1 testing, because it was the key to
part 2 - but I just manually plotted the points. My mistake. Still, it
was nice to check the visualization function against part 1 once I knew
it was right. So it all ended quite well.