Finished Day 14
This was fun as the simple approach for part 1 wasn't going to fly for
part 2, and it was clear that a simplification needed to be made. The
trick was that *all* pairs are expanded the same, and so we can count up
the number of occurrences of each pair, and then as with the octopi, we
can exmapnd them as a population. The trick in counting then, was to
only look at the last letter of each pair, after the 40 steps, and then
add in the very first character in the template - as it's a constant.
This was very fun, but it did take some thinking on the simplification
and the final counting.