Finished Day 16
This was fun because it's all in the data structures. By seeing that I
needed to return the unparsed bits, it made the rest of the day so much
easier. Then it's just a simple process of decoding the bits, and then
making the function to carry out all the operations in the packets.