Finished Day 17
This was going to be an intelligent scanning of the velocity vectors,
but that was foiled by the fact that the probe had to land on the
target, as opposed to pass through it. That meant it wasn't about
aiming, as much as an exhaustive search, and that held true for the
second part too. Not as fun, but easy when you know that's all it is.