Finished Day 18
This was a lot of detailed work to get to the proper exploding and
splitting - as you had to count the nesting, and then regexes were your
best friend after that. But getting them not too greedy was the key.
Then it was the same with regexes and the magnitude, and then it was
just brute force. Not bad, but more time than I expected... and it was
OK.