Finished Day 19
This was exceptionally nasty... the rotations were tricky, but I could
visualize them as combinations of translation functions, and that
helped, and then the cost of computing the brute force approach. So I
had to come up with a way to make it all tractable - and that meant not
only finding out which scans matched - by matching the distances of the
beacons within a set, but then also order the mapping of the
translations so that we didn't have to search for anything. It was all
right there. Not fun - but done. Took me all day.