Finished with Day 20
This was interesting in that the background of the image - on it's
infinite space, would flip from dark to light and back. So the initial
attempts to assume a dark void were off... and so I had to figure out
how to determine the next void from the current void and the enhancement
algo. But in the end, not bad.