Finished Day 21
This was not bad for the first one, but I needed to read up on some
schemes to make the second one a tractable solution. What I also learned
was the Javascript can be much faster than Clojure in these cases, and
that using reduce for the recursion was an interesting way to solve
the problem.