Finished Day 7
This was a parsing and state management problem, and it really came down
to the directories being nested within one another, so that was the
"trick". In the end, I needed to have the entire path as the key in the
size map, but that was then easy. All done.