Finished Day 10
I don't know why I enjoy these "CRT" days so much, but the idea of
creating this problem is just so amazing to me, that it always makes me
giggle. The only trick today was the idea of making the sequence easy
to work with by offsetting it by the initial value of X, do the position
in the sequence was the "middle of the cycle" value. Then it was easy.