Finished Day 11
This was a guess on the last part, confirned by the reddit folks - it
seemed only reasonable that the "rollover" value was related to the fact
that all tests were mobulo operators, so the product of all of them was
the obvious choice, but I wasn't sure. Turns out it was a good guess.