Finished Day 14
This was interesting in that part 2 was pretty much a computationally
intensive problem as there is no way to predict a pattern to when the
sand would fill up the cave. Still... it wasn't too bad, I just had to
make sure the falling sand worked right in both cases.